--- src/components/atoms/baseColors.ts
@@ -1,18 +1,0 @@
-export enum BaseColors {
-	WHITE = 'white',
-	CREAM = 'cream',
-	RED = 'red',
-	DARK = 'dark',
-	SALMON = 'salmon',
-	LIGHT_TONE = 'lightTone',
-	MID_TONE = 'midTone',
-	BIBLE_B = 'bibleB',
-	BOOK_B = 'bookB',
-	BOOK_H = 'bookH',
-	SONG_B = 'songB',
-	SONG_H = 'songH',
-	STORY_B = 'storyB',
-	STORY_H = 'storyH',
-	TOPIC_B = 'topicB',
-	SMART_PLAYLIST_H = 'smartPlaylistH',
-}
