+++ src/components/atoms/horizontalRule.tsx
@@ -1,7 +1,8 @@
 import clsx from 'clsx';
 import React from 'react';
 
-import { BaseColors } from './baseColors';
+import { BaseColors } from '@lib/constants';
+
 import baseColors from './baseColors.module.scss';
 import styles from './horizontalRule.module.scss';
