+++ src/components/molecules/buttonPlay.tsx
@@ -2,7 +2,7 @@ import clsx from 'clsx';
 import React from 'react';
 import { useIntl } from 'react-intl';
 
-import { BaseColors } from '@components/atoms/baseColors';
+import { BaseColors } from '@lib/constants';
 import { ButtonPlayFragment } from '@lib/generated/graphql';
 import usePlaybackSession from '@lib/usePlaybackSession';
