+++ src/components/molecules/card/audiobookTrack.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { FormattedMessage, useIntl } from 'react-intl';
 
-import { BaseColors } from '@components/atoms/baseColors';
 import Heading2 from '@components/atoms/heading2';
 import Heading6 from '@components/atoms/heading6';
 import CardWithPlayable from '@components/molecules/card/base/withPlayable';
+import { BaseColors } from '@lib/constants';
 import { CardAudiobookTrackFragment } from '@lib/generated/graphql';
 
 import HatIcon from '../../../../public/img/fa-book.svg';
