+++ src/components/molecules/collectionTypeLockup.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
 
-import { BaseColors } from '@components/atoms/baseColors';
+import { BaseColors } from '@lib/constants';
 
 import ListIcon from '../../../public/img/fa-list.svg';
