+++ src/components/molecules/definitionList.tsx
@@ -1,8 +1,8 @@
 import clsx from 'clsx';
 import React from 'react';
 
-import { BaseColors } from '@components/atoms/baseColors';
 import Heading6 from '@components/atoms/heading6';
+import { BaseColors } from '@lib/constants';
 
 import baseColorStyles from '../atoms/baseColors.module.scss';
