+++ src/components/molecules/namedAvatar.tsx
@@ -1,8 +1,8 @@
 import clsx from 'clsx';
 import React from 'react';
 
-import { BaseColors } from '@components/atoms/baseColors';
 import RoundImage from '@components/atoms/roundImage';
+import { BaseColors } from '@lib/constants';
 
 import baseColorStyles from '../atoms/baseColors.module.scss';
