+++ src/components/molecules/sponsorTypeLockup.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
 
-import { BaseColors } from '@components/atoms/baseColors';
+import { BaseColors } from '@lib/constants';
 
 import UserPlusIcon from '../../../public/img/fa-user-plus.svg';
