+++ src/components/molecules/typeLockup.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { BaseColors } from '@components/atoms/baseColors';
 import Heading6 from '@components/atoms/heading6';
+import { BaseColors } from '@lib/constants';
 
 import baseColorStyles from '../atoms/baseColors.module.scss';
