+++ src/components/organisms/sequence.tsx
@@ -2,7 +2,6 @@ import Link from 'next/link';
 import React from 'react';
 import { FormattedMessage, useIntl } from 'react-intl';
 
-import { BaseColors } from '@components/atoms/baseColors';
 import Heading2 from '@components/atoms/heading2';
 import Heading6 from '@components/atoms/heading6';
 import HorizontalRule from '@components/atoms/horizontalRule';
@@ -16,6 +15,7 @@ import IconButton from '@components/molecules/iconButton';
 import Tease from '@components/molecules/tease';
 import TeaseHeader from '@components/molecules/teaseHeader';
 import TypeLockup from '@components/molecules/typeLockup';
+import { BaseColors } from '@lib/constants';
 import { formatDateRange } from '@lib/date';
 import { SequenceFragment } from '@lib/generated/graphql';
 import { useFormattedDuration } from '@lib/time';
