+++ src/containers/blog/detail.tsx
@@ -2,12 +2,12 @@ import Image from 'next/image';
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
 
-import { BaseColors } from '@components/atoms/baseColors';
 import HorizontalRule from '@components/atoms/horizontalRule';
 import LineHeading from '@components/atoms/lineHeading';
 import withFailStates from '@components/HOCs/withFailStates';
 import CardColumn from '@components/molecules/card/column';
 import CardPost from '@components/molecules/card/post';
+import { BaseColors } from '@lib/constants';
 import { formatLongDate } from '@lib/date';
 import { GetBlogDetailDataQuery } from '@lib/generated/graphql';
 import { useFormattedDuration } from '@lib/time';
